Refactored and cleaned up
