--- AWSmacOSSDKBuilder/AppDelegate.swift
@@ -1,26 +1,0 @@
-//
-//  AppDelegate.swift
-//  AWSmacOSSDKBuilder
-//
-//  Created by Wayne on 2/17/18.
-//  Copyright © 2018 Amazon Web Services. All rights reserved.
-//
-
-import Cocoa
-
-@NSApplicationMain
-class AppDelegate: NSObject, NSApplicationDelegate {
-
-
-
-	func applicationDidFinishLaunching(_ aNotification: Notification) {
-		// Insert code here to initialize your application
-	}
-
-	func applicationWillTerminate(_ aNotification: Notification) {
-		// Insert code here to tear down your application
-	}
-
-
-}
-
