--- AWSmacOSSDKBuilder/ViewController.swift
@@ -1,27 +1,0 @@
-//
-//  ViewController.swift
-//  AWSmacOSSDKBuilder
-//
-//  Created by Wayne on 2/17/18.
-//  Copyright © 2018 Amazon Web Services. All rights reserved.
-//
-
-import Cocoa
-
-class ViewController: NSViewController {
-
-	override func viewDidLoad() {
-		super.viewDidLoad()
-
-		// Do any additional setup after loading the view.
-	}
-
-	override var representedObject: Any? {
-		didSet {
-		// Update the view, if already loaded.
-		}
-	}
-
-
-}
-
