+++ AWSmacOSTestHost/AWSmacOSTestHost.entitlements
