+++ AWSmacOSTestHost/Assets.xcassets/AppIcon.appiconset/Contents.json
