+++ AWSmacOSTestHost/Info.plist
